Formulario de registro de doctores funcional
* Se agrega el formulario de registro de doctores de manera completamente funcional.
* Se añaden validaciones para el formulario de registro de doctores.

+++ lib/src/domain/entities/patient.dart
@@ -15,7 +15,7 @@ class Patient extends Person {
   String gender;
   String eps;
   int stratum;
-  int discount;
+  double discount;
 
   Patient({
     this.id,
